Move several version routes to version::deprecated::*
